Refactor formatting, change FormatAmount to keep decimal in some cases
When in 7-digit precision mode and formatting adds a thousands
separator, do not truncate all zeroes in case of 7-zero fractional part,
but keep two zeroes instead.
