+++ go/stellar/stellar.go
@@ -1,7 +1,6 @@
 package stellar
 
 import (
-	"bytes"
 	"context"
 	"errors"
 	"fmt"
@@ -882,117 +881,6 @@ func identifyRecipient(m libkb.MetaContext, assertion string, isCLI bool) (keyba
 	frep.Breaks = resp.TrackBreaks
 
 	return frep, nil
-}
-
-func FormatCurrency(ctx context.Context, g *libkb.GlobalContext, amount string, code stellar1.OutsideCurrencyCode) (string, error) {
-	conf, err := g.GetStellar().GetServerDefinitions(ctx)
-	if err != nil {
-		return "", err
-	}
-	currency, ok := conf.Currencies[code]
-	if !ok {
-		return "", fmt.Errorf("FormatCurrency error: cannot find curency code %q", code)
-	}
-
-	amountFmt, err := FormatAmount(amount, true)
-	if err != nil {
-		return "", err
-	}
-
-	if currency.Symbol.Postfix {
-		return fmt.Sprintf("%s %s", amountFmt, currency.Symbol.Symbol), nil
-	}
-
-	return fmt.Sprintf("%s%s", currency.Symbol.Symbol, amountFmt), nil
-}
-
-func FormatCurrencyLabel(ctx context.Context, g *libkb.GlobalContext, code stellar1.OutsideCurrencyCode) (string, error) {
-	conf, err := g.GetStellar().GetServerDefinitions(ctx)
-	if err != nil {
-		return "", err
-	}
-	currency, ok := conf.Currencies[code]
-	if !ok {
-		return "", fmt.Errorf("FormatCurrencyLabel error: cannot find curency code %q", code)
-	}
-	return fmt.Sprintf("%s (%s)", code, currency.Symbol.Symbol), nil
-}
-
-func FormatPaymentAmountXLM(amount string, delta stellar1.BalanceDelta) (string, error) {
-	desc, err := FormatAmountXLM(amount)
-	if err != nil {
-		return "", err
-	}
-
-	switch delta {
-	case stellar1.BalanceDelta_DECREASE:
-		desc = "- " + desc
-	case stellar1.BalanceDelta_INCREASE:
-		desc = "+ " + desc
-	}
-
-	return desc, nil
-}
-
-// Example: "157.5000000 XLM"
-func FormatAmountXLM(amount string) (string, error) {
-	return FormatAmountWithSuffix(amount, false, "XLM")
-}
-
-func FormatAmountWithSuffix(amount string, precisionTwo bool, suffix string) (string, error) {
-	formatted, err := FormatAmount(amount, precisionTwo)
-	if err != nil {
-		return "", err
-	}
-	return fmt.Sprintf("%s %s", formatted, suffix), nil
-}
-
-func FormatAmount(amount string, precisionTwo bool) (string, error) {
-	if amount == "" {
-		return "", errors.New("empty amount")
-	}
-	x, err := stellarnet.ParseDecimalStrict(amount)
-	if err != nil {
-		return "", fmt.Errorf("unable to parse amount %s: %v", amount, err)
-	}
-	precision := 7
-	if precisionTwo {
-		precision = 2
-	}
-	s := x.FloatString(precision)
-	parts := strings.Split(s, ".")
-	if len(parts) != 2 {
-		return "", fmt.Errorf("unable to parse amount %s", amount)
-	}
-	if parts[1] == "0000000" {
-		// get rid of all zeros after point if default precision
-		parts = parts[:1]
-	}
-	head := parts[0]
-	if len(head) <= 3 {
-		return strings.Join(parts, "."), nil
-	}
-	sinceComma := 0
-	var b bytes.Buffer
-	for i := len(head) - 1; i >= 0; i-- {
-		if sinceComma == 3 && head[i] != '-' {
-			b.WriteByte(',')
-			sinceComma = 0
-		}
-		b.WriteByte(head[i])
-		sinceComma++
-	}
-	parts[0] = reverse(b.String())
-
-	return strings.Join(parts, "."), nil
-}
-
-func reverse(s string) string {
-	r := []rune(s)
-	for i, j := 0, len(r)-1; i < len(r)/2; i, j = i+1, j-1 {
-		r[i], r[j] = r[j], r[i]
-	}
-	return string(r)
 }
 
 // ChangeAccountName changes the name of an account.
